adding Hash Table
I noticed there is a file reading function that involves the binary
search tree. However, my technique of getting data into hash table is
different, but I dare not to modify the function cause that might be
affect the search tree, so I created a new reading function right above
that function.