Hash Table correction 1
I cannot compile the program, so I spent some time to fix the
problems...
This is a temporary fix for my hash table. I disable the search tree as
it doesn't work...
Somebody has deleted #includes in "HashTable.cpp" and "RunOption.cpp". I
have added it back.
Also, for my getData function in RunOption, please do not change the
return type to bool becuase the pointer of hashtable will die.
Actually, we don't have to pass the hashtable pointer as argument
because we have defined hashTable pointer in header file, same as search
trees. Therefore, perhaps we can have a readfile function that put a
website in trees and hash table at the same time.